requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
packaging==25.0
pluggy==1.5.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.1.0
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
flake8==6.1.0
openai==1.13.3
httpx==0.27.0
requests>=2.0
markdown
gunicorn
